Move parseVHF into VHF
Moves relevant import spaces too
