+++ Archive/20230901 Threading Speed/test-thread-filename.py
@@ -37,8 +37,8 @@ from pathlib import Path
 module_path = str(Path(__file__).parents[2])
 if module_path not in sys.path:
     sys.path.append(module_path)
+from VHF.parse import VHFparser
 from VHF.runner import VHFRunner
-from parseVHF import VHFparser
 
 
 _PATH = Union[str, bytes, PathLike, Path]
